+++ web/static/js/app.js
@@ -18,4 +18,4 @@ import "phoenix_html"
 // Local files can be imported directly using relative
 // paths "./socket" or full ones "web/static/js/socket".
 
-import socket from "./socket"
+import Kafkamon from "./kafkamon/components/kafkamon"
