+++ web/static/js/sidebar/components/topic_link.jsx
@@ -1,0 +1,12 @@
+class TopicLink extends React.Component {
+  handleClick() {
+    this.props.onClick(this.props.name)
+  }
+
+  render(){
+    return(
+      <a style={{cursor: "pointer"}} onClick={() => this.handleClick()}>{this.props.name}</a>
+    )
+  }
+}
+export default TopicLink
